Refactored SerializationArchive
now supports multiple data providers and checksum may be applied to each,
also data is lazy loaded on demand
