+++ framework/serialization/serialize.h
@@ -2,7 +2,6 @@
 
 #include "library/sp.h"
 #include "library/strings.h"
-
 #include <exception>
 #include <iterator>
 #include <list>
